ABSPATH
1. defined( 'ABSPATH' )

The defined() function checks whether a constant named 'ABSPATH' is defined or not.
'ABSPATH' is a constant used in WordPress to indicate the absolute path to the WordPress installation. This constant is typically defined in the wp-config.php file.

2. !

This symbol means “not.” Therefore, ! defined( 'ABSPATH' ) means that if 'ABSPATH' is not defined, the code inside the if block will be executed.

3. exit;

This statement causes the script to terminate. In other words, if 'ABSPATH' is not defined, the script stops, and no code after this line will be executed.